refactored preference encoding and decoding. Also added comments to explain the functions
